+++ tests/BusinessLogic/SeQuraAPI/Order/OrderProxyTest.php
@@ -946,7 +946,6 @@ class OrderProxyTest extends BaseTestCase
             'testRejectCallback',
             'testPartPaymentDetailsGetter',
             'testApprovedUrl',
-            'testOperatorRef',
             new Options(true,false,true,true),
             new EventsWebhook('https://testUrl', ['signature' => 'testSignature', 'testParam1Key' => 'testParam1Value'])
         );
@@ -1174,7 +1173,6 @@ class OrderProxyTest extends BaseTestCase
             'testRejectCallback',
             'testPartPaymentDetailsGetter',
             'testApprovedUrl',
-            'testOperatorRef',
             new Options(true,false,true,true),
             new EventsWebhook('https://testUrl', ['signature' => 'testSignature', 'testParam1Key' => 'testParam1Value'])
         );